Restructure team members
