+++ pkg/registry/apis/identity/legacy/queries.go
@@ -4,9 +4,6 @@ import (
 	"embed"
 	"fmt"
 	"text/template"
-
-	"github.com/grafana/grafana/pkg/storage/legacysql"
-	"github.com/grafana/grafana/pkg/storage/unified/sql/sqltemplate"
 )
 
 // Templates setup.
@@ -22,31 +19,4 @@ func mustTemplate(filename string) *template.Template {
 		return t
 	}
 	panic(fmt.Sprintf("template file not found: %s", filename))
-}
-
-// Templates.
-var (
-	sqlQueryTeamMembers = mustTemplate("query_team_members.sql")
-)
-
-type sqlQueryListTeamMembers struct {
-	sqltemplate.SQLTemplate
-	Query           *ListTeamMembersQuery
-	UserTable       string
-	TeamTable       string
-	TeamMemberTable string
-}
-
-func (r sqlQueryListTeamMembers) Validate() error {
-	return nil // TODO
-}
-
-func newListTeamMembers(sql *legacysql.LegacyDatabaseHelper, q *ListTeamMembersQuery) sqlQueryListTeamMembers {
-	return sqlQueryListTeamMembers{
-		SQLTemplate:     sqltemplate.New(sql.DialectForDriver()),
-		UserTable:       sql.Table("user"),
-		TeamTable:       sql.Table("team"),
-		TeamMemberTable: sql.Table("team_member"),
-		Query:           q,
-	}
 }
