--- pkg/registry/apis/identity/legacy/query_users.sql
@@ -1,13 +1,0 @@
-SELECT o.org_id, u.id, u.uid, u.login, u.email, u.name,
-  u.created, u.updated, u.is_service_account, u.is_disabled, u.is_admin
-  FROM {{ .Ident .UserTable }} as u JOIN {{ .Ident .OrgUserTable }} as o ON u.id = o.user_id
- WHERE o.org_id = {{ .Arg .Query.OrgID }}
-   AND u.is_service_account = {{ .Arg .Query.IsServiceAccount }}
-{{ if .Query.UID }}
-   AND u.uid = {{ .Arg .Query.UID }}
-{{ end }}
-{{ if .Query.Pagination.Continue }}
-   AND u.id >= {{ .Arg .Query.Pagination.Continue }}
-{{ end }}
- ORDER BY u.id asc
- LIMIT {{ .Arg .Query.Pagination.Limit }}
