+++ pkg/registry/apis/identity/legacy/types.go
@@ -4,20 +4,8 @@ import (
 	"context"
 
 	"github.com/grafana/authlib/claims"
-	"github.com/grafana/grafana/pkg/registry/apis/identity/common"
 	"github.com/grafana/grafana/pkg/services/team"
 )
-
-type ListTeamMembersQuery struct {
-	UID        string
-	OrgID      int64
-	Pagination common.Pagination
-}
-
-type ListTeamMembersResult struct {
-	Continue int64
-	Members  []TeamMember
-}
 
 // In every case, RBAC should be applied before calling, or before returning results to the requester
 type LegacyIdentityStore interface {
